my_site/requirements.txt
Django==5.2.3
daphne==3.0.2
Unidecode==1.4.0

python-dotenv==1.0.1
cloudinary==1.42.1
whitenoise==6.8.2

pillow==11.1.0
django-cloudinary-storage==0.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
django-modeltranslation==0.19.12
idna==3.10
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0


channels==3.0.5
channels-redis==4.1.0
redis==6.2.0

requests==2.32.3
beautifulsoup4==4.13.4

python-docx>=1.0.0
num2words>=0.5.13
